Crate terminal_graphics

Terminal Graphics

A small package for rust which could do wonders when in right hands :) For rendering things on terminal

Structs

BarCounter

Make a BarCounter which can be updated

DiamondObject

Diamond shaped Obj which can be rendered on the terminal

LineCounter

Makes a LineCounter which can be updated

RectangularObject

Rectangular Obj which can be rendered on the terminal

Functions

animated_line

Animated line

bar_animation

Make a bar animation

clear_terminal

Clears the terminal for you

get_terminal_size

Get the size of the terminal

new_row

Jump into the next row

print

Prints a String

print_backspace

Makes a backspace in terminal

print_color

Print a string with custom Background color

print_dash

Prints a line

print_error

Print Error

print_fake

Prints a String which can be replaced

print_info

Print Info

print_pixel

Print a pixel on the terminal

print_pixels

Print a number of pixels

print_success

Print Success

print_tab

Prints a tab

print_warning

Print warning

println

Prints a String along with a newline

println_color

Print a string with custom Background color and newline

remove_pixel

Remove a pixel from the current row

remove_pixels

Remove a number of pixels from current row

set_cursor

Set the cursor

spin

Makes a spin

spinner

Spinner